% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{add.lagged.covariates}
\alias{add.lagged.covariates}
\title{Augment dataframe with lagged covariates}
\usage{
add.lagged.covariates(dat, outcomename, covariates = NULL)
}
\arguments{
\item{dat}{The dataframe}

\item{outcomename}{The outcome of interest (string)}

\item{covariates}{The covariates to lag (list of string names).  NULL if no
covariates other than outcome should be lagged.  covariates can also be a
function with a "lags" attribute with the listed covariates (as returned
by, e.g., make.fit.season.model)}
}
\value{
Augmented dataframe
}
\description{
Augment dataframe with lagged covariates
}
