% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratified_ITS.R
\name{adjust.data}
\alias{adjust.data}
\title{Adjust an outcome based on the group weights.}
\usage{
adjust.data(dat, outcomename, groupname, pi.star, is.count = FALSE,
  include.aggregate = FALSE, covariates = NULL)
}
\arguments{
\item{dat}{Dataframe of data.  Requires an N column of total cases
represented in each row.}

\item{outcomename}{Name of column that has the outcome to calculated adjusted
values for.}

\item{groupname}{Name of categorical covariate that determines the groups.}

\item{pi.star}{The target weights.  Each month will have its groups
re-weighted to match these target weights.}

\item{include.aggregate}{Include aggregated (unadjusted) totals in the output
as well.}
}
\description{
Adjust an outcome based on the group weights.
}
