% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{aggregate.simulation.results}
\alias{aggregate.simulation.results}
\title{Test a passed test statistic on the simulated data}
\usage{
\method{aggregate}{simulation.results}(orig.data, predictions, outcomename,
  summarizer = calculate.average.outcome, ...)
}
\arguments{
\item{orig.data}{The raw data (dataframe)}

\item{predictions}{The results from process.outcome.model}

\item{outcomename}{Outcome to use}

\item{summarizer}{A function to calculate some summary quantity, Default: calculate.average.outcome}

\item{...}{Arguments to be passed to summarizer()}
}
\value{
List of the test statistic and reference distribution.
}
\description{
This method is used to look at summary statistics such as average impact post-policy, and see how the predictive distribution compares to the observed.
}
