% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratified_ITS.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate grouped data}
\usage{
aggregate_data(
  dat,
  outcomename,
  groupname,
  is_count = FALSE,
  rich = TRUE,
  covariates = NULL
)
}
\arguments{
\item{dat}{Dataframe of data}

\item{outcomename}{String name of the outcome variable in dat.}

\item{groupname}{Name of grouping variable to post-stratify on.}

\item{is_count}{If TRUE the data are counts, and should be aggregated by sum
rather than by mean.}

\item{rich}{If TRUE, add a bunch of extra columns with proportions of the
month that are each group and so forth.}

\item{covariates}{group-invariant covariates to preserve in the augmented
rich dataframe.  These are not used in this method for any calculations.
Pass as list of column names of dat}
}
\description{
This will take a dataframe with each row being the outcomes, etc., for a
given group for a given month and aggregate those groups for each month.
}
