% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{aggregate_simulation_results}
\alias{aggregate_simulation_results}
\title{Test a passed test statistic on the simulated data}
\usage{
aggregate_simulation_results(
  orig.data,
  predictions,
  outcomename,
  summarizer = calculate.average.outcome,
  ...
)
}
\arguments{
\item{orig.data}{The raw data (dataframe)}

\item{predictions}{The results from process.outcome.model.}

\item{outcomename}{Outcome to use.}

\item{summarizer}{A function to calculate some summary quantity, Default:
calculate.average.outcome}

\item{...}{Extra arguments passed to the summarizer function.}
}
\value{
List of the test statistic and reference distribution.
}
\description{
This method is used to look at summary statistics such as
  average impact post-policy, and see how the predictive distribution
  compares to the observed.
}
