% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{calculate.average.outcome}
\alias{calculate.average.outcome}
\title{Summary function for summarize.simulation.results}
\usage{
calculate.average.outcome(res, outcomename, months = 1:18, ...)
}
\arguments{
\item{res}{Dataframe of a single series (simulated or otherwise)}

\item{outcomename}{Name of outcome in res}

\item{months}{Which months to average over, Default: 1:18}

\item{...}{Other parameters (ignored)}
}
\value{
Single number (in this case mean of given months)
}
\description{
Given a set of simulation runs, estimate average impact over range of months
}
\examples{
data( mecklenberg )
calculate.average.outcome( mecklenberg, "pbail" )
}
