% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratified_ITS.R
\name{calculate.group.weights}
\alias{calculate.group.weights}
\title{Calculate overall proportion of cases in each group that lie within a given
interval of time defined by t.min and t.max.}
\usage{
calculate.group.weights(groupname, dat, t.min, t.max = max(dat$month), N = "N")
}
\arguments{
\item{groupname}{Name of the column that has the grouping categorical
variable}

\item{dat}{Dataframe with one row for each time point and group that we are
going to post stratify on.  This dataframe should also have an 'N' column
indicating the number of cases that make up each given row. It should have
a 'month' column for the time.}

\item{t.min}{The start month to aggregate cases over.}

\item{t.max}{The final month (default is last month).}
}
\description{
Calculate overall proportion of cases in each group that lie within a given
interval of time defined by t.min and t.max.
}
