% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_model_functions.R
\name{fit_model_default}
\alias{fit_model_default}
\title{Default ITS model}
\usage{
fit_model_default(dat, outcomename, lagless = FALSE, ...)
}
\arguments{
\item{dat}{Dataframe of pre-policy data to fit model to.  Needs a "month" column}

\item{outcomename}{Outcome of interest}

\item{lagless}{Boolean, include the lagged outcome, or not?}

\item{...}{Extra arguments passed to the lm() call.}
}
\description{
This fits the model outcomename ~ lag.outcome + month, with no
covariates.
}
