% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{generate.prediction.sequence}
\alias{generate.prediction.sequence}
\title{Given a model captured in beta.vec (the vector of coefficients from a
regression held in 'fit0') and sigma (the standard deviation of the
residuals), make a simulated series of time points that follow the model.}
\usage{
generate.prediction.sequence(beta.vec, sigma, dat, fit0, outcomename, t0)
}
\arguments{
\item{beta.vec}{Vector of parameter values}

\item{sigma}{Value for standard deviation of independent portion of residual}

\item{dat}{The dataframe (pre and post policy)}

\item{fit0}{The original model fit to the data}

\item{outcomename}{The outcome of interest}

\item{t0}{The last pre-policy timepoint.}
}
\value{
Dataframe, one row per timepoint.  Columns are 'month' (the time),
  'Ybar' the predicted outcomes from the model, and 'Ystar', a concatenation
  of the pre-policy series followed by the synthetic sequence.
}
\description{
Given a model captured in beta.vec (the vector of coefficients from a
regression held in 'fit0') and sigma (the standard deviation of the
residuals), make a simulated series of time points that follow the model.
}
