% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{generate_structure_sequence}
\alias{generate_structure_sequence}
\title{Generate the predictions of fit0 excluding the lagged outcome variable.}
\usage{
generate_structure_sequence(fit0, dat)
}
\arguments{
\item{fit0}{Model fit to data.}

\item{dat}{Data to predict outcomes for using fit0.}
}
\value{
Predictions as a numeric vector.
}
\description{
This is a utility function used by the make_autoregressive_series() method.
}
\details{
We do this so we can subtract this part off before smoothing so our smoother
doesn't have to try and capture the seasonality stuff.
}
