% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{generate_Ybars}
\alias{generate_Ybars}
\title{Generate the model predictions for a series, ignoring lagged components.}
\usage{
generate_Ybars(fit_model, outcomename, t0, dat)
}
\arguments{
\item{fit_model}{A function that takes data, fits a linear model, and returns
the fit model. This function needs an option to include (or not) lagged
covariates.}

\item{outcomename}{String name of the outcome variable in dat.}

\item{t0}{last pre-policy timepoint}

\item{dat}{Dataframe of the data.}
}
\value{
Predicted sequence for passed dataframe (as vector)
}
\description{
Generate Ybar given a dataframe 'dat'. Ybar is our model predictions without
any lagged terms, fit to the pre-policy data only. This is basically the
"structural" component of our model, which can include seasonality.
}
\details{
This is so we can show trendlines and so forth on plots.

Ybar is defined for both pre and post policy datapoints.
}
