% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{make.autoregressive.series}
\alias{make.autoregressive.series}
\title{Given model parameters, make a synthetic timeseries for all points after t0.
This series is a plausible continuation of a pre-policy series.  This method needs
to be given the structural component, calculated from the intercept, linear trend,
temperature, and seasonality.
Bundling all seasonality, etc., into Ybar, the assumed model is:
   $$Y = Ybar + beta1 * Y.pre + residual$$
where Ybar is our structural model}
\usage{
make.autoregressive.series(Ybar, Y.init, beta1, sigma, expected = FALSE,
  include.start = FALSE)
}
\arguments{
\item{Ybar}{Structural (expected) sequence.  This method adds the autoregressive residual component
to this.}

\item{Y.init}{Value of series at time point just before first time point in dat.post}

\item{beta1}{Coefficient for lagged outcome}

\item{sigma}{Standard deviation of the residual errors to add to each step}

\item{expected}{If TRUE, return the expected value of the series with no residual noise}

\item{include.start}{If TRUE, return Y.init as the initial element in the returned sequence.}
}
\value{
Vector of the synthetic outcomes.
}
\description{
Given model parameters, make a synthetic timeseries for all points after t0.
This series is a plausible continuation of a pre-policy series.  This method needs
to be given the structural component, calculated from the intercept, linear trend,
temperature, and seasonality.
Bundling all seasonality, etc., into Ybar, the assumed model is:
   $$Y = Ybar + beta1 * Y.pre + residual$$
where Ybar is our structural model
}
