% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{make.envelope.graph}
\alias{make.envelope.graph}
\title{Make envelope style graph with associated smoothed trendlines}
\usage{
make.envelope.graph(envelope, t0, ylab = "Y", xlab = "month")
}
\arguments{
\item{envelope}{The result of a process.outcome.model call, i.e. dataframe
with columns of original data, imputed data and, potentially, smoothed
data.}

\item{t0}{Last pre-policy timepoint.}

\item{ylab}{Y label of plot}

\item{xlab}{X label of plot}
}
\description{
Make envelope style graph with associated smoothed trendlines
}
