% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_code.R
\name{make.fake.data}
\alias{make.fake.data}
\title{Make fake data for testing purposes.}
\usage{
make.fake.data(
  t.min = -40,
  t.max = 9,
  t0 = 0,
  rho = 0.5,
  sd.omega = 1,
  coef.line = c(20, 0.05),
  coef.q = c(1, 0, -1, 0),
  coef.temp = 0.1,
  coef.sin = c(0, 0),
  coef.tx = c(0, 0.25, 5)
)
}
\arguments{
\item{t.min}{Index of first month}

\item{t.max}{Index of last month}

\item{t0}{Last pre-policy time point}

\item{rho}{Autocorrelation}

\item{sd.omega}{Standard deviation of the true residual}

\item{coef.line}{Intercept and slope of the main trendline (list of 2).}

\item{coef.q}{Coefficients for the four quarters (list of 4).}

\item{coef.temp}{Coefficient for temperature.}

\item{coef.sin}{Coefficents for sin and cos features (list of 2)}

\item{coef.tx}{Coefficient for treatment post-policy (list of 3, initial
offset, initial slope, additional slope past 12 months).  Treatment is a
piecewise linear function.}
}
\value{
A \code{data.frame} having \code{month} , \code{temperature} ,
  \code{sin.m} , \code{cos.m} , \code{Q1}, \code{Q2} , \code{Q3}, \code{Q4},
  \code{post} , \code{Ystr0} , \code{Ystr} , \code{Y}
}
\description{
Has heavy seasonality, and an extra bump at 12 months post-policy
}
\examples{
simData <- make.fake.data()
t0 <- 0
simData <- make.fake.data(t.min=-40, t.max=15, t0=t0)

}
