% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratified_ITS.R
\name{make.fake.group.data}
\alias{make.fake.group.data}
\title{A fake DGP for illustrating the code.}
\usage{
make.fake.group.data(
  t.min,
  t0,
  t.max,
  method = c("complex", "linear", "jersey")
)
}
\arguments{
\item{t.min}{Index of first month}

\item{t0}{last pre-policy timepoint}

\item{t.max}{Index of last month}

\item{method}{Type of post-stratification structure to generate.}
}
\description{
This code makes synthetic grouped data that can be used to illustrate
benefits of post stratification.
}
