% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{make.fit.season.model}
\alias{make.fit.season.model}
\title{Make a fit.model that takes a seasonality component}
\usage{
make.fit.season.model(formula, no.lag = NULL)
}
\arguments{
\item{formula}{Formula specifying seasonality.  No outcome or month needed.}

\item{no.lag}{Formula specifying additional variables to not lag (usually due to colinearity).}
}
\value{
A function that takes dat, outcomename, and a lagless flag (see,
  e.g., fit.model.default)
}
\description{
This method returns a function that will fit a model both with and without
lagged outcomes.
}
\details{
You hand it a formula object specifying the seasonality, e.g., " ~ Q2 + Q3 +
Q4", if you have quarterly season effects, and it will assume you want models
with a linear month component as well.
}
