% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{make.many.predictions}
\alias{make.many.predictions}
\title{OLD
#' @return An updated version of the 'res' dataframe with `Ysmooth`, the
  smoothed predictions of the original Ystar outcome.  Also includes 'Ystar'
  the original sequence to be smoothed.  If full.output=TRUE also include
  several other columns: `resid`, the residuals based on Ystar and the model,
  `residStar` the smoothed residuals, 'Ybar.sm' the structural predictions of
  the model used for smoothing.
Generate a collection of raw counterfactual trajectories}
\usage{
make.many.predictions(fit0, dat, R, outcomename, t0)
}
\arguments{
\item{fit0}{The fit linear model to simulate from.}

\item{dat}{A dataframe with the covariates needed by the model fit0 for both pre
and post-policy months.}

\item{R}{Number of series to generate.}

\item{outcomename}{The name of the column in dat which is our outcome.}

\item{t0}{Last month of pre-policy.  Will start predicting at t0+1.}
}
\value{
data.frame with the collection of predicted series
}
\description{
Given a fit linear model 'fit0', generate R prediction series starting at t0.
This takes model uncertainty into account by pulling from the
pseudo-posterior of the model parameters (from Gelman and Hill arm package).
}
