% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{make_fit_season_model}
\alias{make_fit_season_model}
\title{Make a fit_model that takes a seasonality component}
\usage{
make_fit_season_model(formula, no.lag = NULL)
}
\arguments{
\item{formula}{Formula specifying seasonality.  No outcome or month needed.}

\item{no.lag}{Formula specifying additional variables to not lag (usually used due
to colinearity of lagged outcomes, such as with a sin and cos component).}
}
\value{
A function that takes dat, outcomename, and a lagless flag (see,
  e.g., fit_model.default)
}
\description{
This method returns a function that will fit a model both with and without
lagged outcomes.
}
\details{
You hand it a formula object specifying the seasonality, e.g., " ~ Q2 + Q3 +
Q4", if you have quarterly season effects. This method assumes you want
models with a linear month component as well, and will add that and an
intercept in automatically.
}
