% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meck_subgroup}
\alias{meck_subgroup}
\title{Mecklenberg subgroup data}
\format{A data frame with 144 rows and 5 variables:
\describe{
  \item{\code{month}}{integer Month, with 0 being month of policy implementation.}
  \item{\code{n.cases}}{integer Number of cases total that month}
  \item{\code{n.bail}}{interger number of bail cases that month}
  \item{\code{pbail}}{double Proportion of cases in subgroup assigned bail} 
  \item{\code{category}}{character Category of group}
}}
\usage{
meck_subgroup
}
\description{
Mecklenberg data of proportion of different categories of cases given bail (by month).
}
\keyword{datasets}
