% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mecklenberg}
\alias{mecklenberg}
\title{Mechlenberg PSA Reform Data}
\format{A data frame with 54 rows and 10 variables:
\describe{
  \item{\code{month}}{integer Month, with 0 being month of policy implementation.}
  \item{\code{karr}}{integer COLUMN_DESCRIPTION}
  \item{\code{pbail}}{double Proportion of cases in a given month assigned bail (or outright detention).}
  \item{\code{pptrel}}{double COLUMN_DESCRIPTION}
  \item{\code{pror}}{double COLUMN_DESCRIPTION}
  \item{\code{pb4c}}{double COLUMN_DESCRIPTION}
  \item{\code{avg_days_initial}}{double COLUMN_DESCRIPTION}
  \item{\code{avg_t2d}}{double COLUMN_DESCRIPTION}
  \item{\code{pstint7}}{double COLUMN_DESCRIPTION}
  \item{\code{pstint30}}{double COLUMN_DESCRIPTION}
}}
\usage{
mecklenberg
}
\description{
Monthly aggregate outcomes of various measures of interest from Mechlenberg. See MDRC Report.
}
\keyword{datasets}
