% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newjersey}
\alias{newjersey}
\title{New Jersey PSA Reform aggregate data}
\format{A data frame with 106 rows and 11 variables:
\describe{
  \item{\code{tsin1}}{double COLUMN_DESCRIPTION}
  \item{\code{tcos1}}{double COLUMN_DESCRIPTION}
  \item{\code{M12}}{integer COLUMN_DESCRIPTION}
  \item{\code{Q1}}{integer COLUMN_DESCRIPTION}
  \item{\code{Q2}}{integer COLUMN_DESCRIPTION}
  \item{\code{Q3}}{integer COLUMN_DESCRIPTION}
  \item{\code{Q4}}{integer COLUMN_DESCRIPTION}
  \item{\code{compwarr.cs}}{double Number of warrant arrests}
  \item{\code{compsumm.cs}}{double Number of summons arrests}
  \item{\code{comptot.cs}}{double Total number of arrests}
  \item{\code{temperature}}{double Average temperature in New Jersey that month.}
}}
\usage{
newjersey
}
\description{
Montly aggregate counts of arrests of different types in New Jersey.
}
\keyword{datasets}
