% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{smooth.residuals}
\alias{smooth.residuals}
\title{Smooth a series by fitting the model to the data, smoothing the residuals,
and then putting the model predictions back.}
\usage{
smooth.residuals(res, t0, fit0, outcomename, covariates,
  fit.model = fit.model.default, smooth_k = SMOOTH_K,
  full.output = FALSE)
}
\arguments{
\item{res}{A dataframe with a Ystar column and all the covariates in the
model, Ystar being the outcome to smooth.}

\item{smooth_k}{A rough proxy for the number of observations to primarily
consider to kernal weight in the neighborhood of each timepoint (this is a
bandwidth, and the loess smoother gets smooth_k / n as a span value).  We
want to smooth with an absolute bandwidth, not one as function of how long
the time series is.}

\item{full.output}{Include residuals, etc., to see how smoother operated.}
}
\value{
An updated version of the 'res' dataframe with `Ysmooth`, the
  smoothed predictions of the original Ystar outcome.  Also includes 'Ystar'
  the original sequence to be smoothed.  If full.output=TRUE also include
  several other columns: `resid`, the residuals based on Ystar and the model,
  `residStar` the smoothed residuals, 'Ybar' the structural predictions of
  the model.
}
\description{
Use loess smoother on complete series of residuals including original data
pre-policy and synthetic data post policy (i.e., smooth the entire plausible
series).
}
