% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{smooth.series}
\alias{smooth.series}
\title{Smooth a series using a static loess smoother}
\usage{
smooth.series(res, outcomename, t0, smooth_k = SMOOTH_K, post.only = TRUE, ...)
}
\arguments{
\item{res}{A dataframe with a month column and an 'outcomename' column (which
is the column that will be smoothed).}

\item{outcomename}{String name of the outcome variable in dat.}

\item{t0}{last pre-policy timepoint}

\item{smooth_k}{A rough proxy for the number of observations to primarily
consider to kernal weight in the neighborhood of each timepoint (this is a
bandwidth, and the loess smoother gets smooth_k / n as a span value).  We
want to smooth with an absolute bandwidth, not one as function of how long
the time series is.}

\item{post.only}{Boolean of smooth post-policy data only, or full sequence.}

\item{...}{Extra arguments (not used in this function).}
}
\value{
An updated version of the 'res' dataframe with `Ysmooth`, the
  smoothed predictions of the original Ystar outcome.  Also includes 'Ystar'
  the original sequence to be smoothed.
}
\description{
Use loess smoother on complete series of residuals including original data
pre-policy and synthetic data post policy (i.e., smooth the entire plausible
series).
}
\details{
This method takes several parameters it does not use, to maintain
compatability with smooth.residuals.
}
