% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{generate.structure.sequence}
\alias{generate.structure.sequence}
\title{Generate the predictions of fit0 excluding the lagged outcome variable. This
is a utility function used by the make.autoregressive.series() method.}
\usage{
generate.structure.sequence(fit0, dat)
}
\description{
We do this so we can subtract this part off before smoothing so our smoother
doesn't have to try and capture the seasonality stuff.
}
