% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_code.R
\name{make.fake.data}
\alias{make.fake.data}
\title{Make fake data for testing purposes.}
\usage{
make.fake.data(t.min = -40, t.max = 9, t0 = 0, rho = 0.5)
}
\description{
Make fake data for testing purposes.
}
