% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{make.model.smoother}
\alias{make.model.smoother}
\title{Make a smoother that fits a model and smooths residuals}
\usage{
make.model.smoother(fit.model, covariates)
}
\value{
a smoother function that can be passed to the smoothing routines.
}
\description{
Make a smoother that fits a model and smooths residuals
}
