% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newjersey}
\alias{newjersey}
\title{New Jersey PSA Reform aggregate data}
\format{
A data frame with 106 rows and 11 variables:
\describe{
  \item{\code{month}}{integer Index of month.}
  \item{\code{sin.m}}{double cos of month number}
  \item{\code{cos.m}}{double sin of month number}
  \item{\code{M12}}{integer Month number}
  \item{\code{Q1}}{integer Indicator of 1st quarter.}
  \item{\code{Q2}}{integer Indicator of 2nd quarter.}
  \item{\code{Q3}}{integer Indicator of 3rd quarter.}
  \item{\code{Q4}}{integer Indicator of 4th quarter.}
  \item{\code{n.warrant}}{double Number of warrant arrests}
  \item{\code{n.summons}}{double Number of summons arrests}
  \item{\code{n}}{double Total number of arrests}
  \item{\code{temperature}}{double Average temperature in New Jersey that month.}
}
}
\usage{
newjersey
}
\description{
Montly aggregate counts of arrests of different types in New Jersey.
}
\keyword{datasets}
