% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITSModelingCode.R
\name{smooth.residuals}
\alias{smooth.residuals}
\title{Smooth residuals after model fit}
\usage{
smooth.residuals(
  res,
  t0,
  outcomename,
  post.only = TRUE,
  smooth_k = SMOOTH_K,
  fit.model = fit.model.default,
  covariates,
  full.output = FALSE
)
}
\arguments{
\item{res}{A dataframe with a month column and 'outcomename' column, which is
the column to be smoothed.}

\item{t0}{Last month of pre-policy.  Will start predicting at t0+1.}

\item{outcomename}{The name of the column in dat which is our outcome.}

\item{post.only}{If TRUE fit model and smooth post-policy only. WHY fit model
on post-policy data only?  Because this will make sure the fixed pre-policy
does not dominate too much?  We are focusing on post-policy so we want a
good fitting model for that so we can get our residuals as "white noise" as
possible before smoothing.}

\item{smooth_k}{A rough proxy for the number of observations to primarily
consider to kernal weight in the neighborhood of each timepoint (this is a
bandwidth, and the loess smoother gets smooth_k / n as a span value).  We
want to smooth with an absolute bandwidth, not one as function of how long
the time series is.}

\item{fit.model}{A function that takes data, fits a linear model, and returns
the fit model. This function needs an option to include (or not) lagged
covariates.}

\item{covariates}{A dataframe with all covariates needed in the model fitting
defined by fit.model.}

\item{full.output}{If TRUE give back pieces for diagnostics of smoothing
process.}
}
\description{
Smooth a series by fitting the model to the data, smoothing the residuals,
and then putting the model predictions back.
}
\details{
Use loess smoother on complete series of residuals including original data
pre-policy and synthetic data post policy (i.e., smooth the entire plausible
series).
}
